Rename files and folders to remove case & spaces
Files and folders with spaces in their names, or that are case sensitive can prove problematic when attempting to work with them from the command line (though admittedly not impossible).

To protect against any possible issues this change renames certain folders and files to remove make them case insensitive and/or remove any spaces.