Add ability to set a custom window size (#63)
This change fixes an issue that now appears to be common when running on Linux as opposed to OSX.

If an element is not visible in the window then capybara can't find it so the test fails. If you run without setting a custom window size, you can get your tests to pass by manually resizing the window hence confirming this is the issue.

With this change you can now set a custom window size, that will be used across all tests.