Lock chromedriver-helper to 1.0.0
The current belief based on our experience with other projects is because the chromedriver version of the project is now no longer in sync with that installed on our Jenkins slave it can no longer connect and run.

This change locks the version of chromedriver-helper to one we know that works in our environments.