Attempt to fix not working in Jenkins CI
The error we are getting in Jenkins is

```
unable to connect to chromedriver 127.0.0.1:9515 (Selenium::WebDriver::Error::WebDriverError)
```

In previous services this has been because the version of chromedriver used by the project has become out of sync with the version on the Jenkins slave.

However for PAFS we checked the versions of Chrome and Chromedriver and found

- Chrome 66.0.3359.181
- Chromedriver 2.38.552522

At the time of writing this would indicate the slave is actually using the latest versions of everything.

Hence this change attempts to fix the issue by moving chromedriver-helper to the latest.